requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
finist==0.1.2
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.11
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-asyncio==0.20.3
python-dotenv==0.21.1
redis==4.5.1
SQLAlchemy==2.0.4
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
yarl==1.8.2
